+++ replay/csv_conversion_utils.py
@@ -1,11 +1,15 @@
-from utils import *
+import utils
+import json
+import math
+import time
+from decoded_messages import DecodedMessage
 
-def csv_conversion(filename: str, csv_filename: str, csv_interpolation: bool, start_time: int, end_time: int, agent_id: int = 1, agent_type: str = "car"):
+def csv_conversion(input_filename: str, csv_filename: str, csv_interpolation: bool, start_time: int, end_time: int, agent_id: int = 1, agent_type: str = "car"):
     """
     CSV function to store in a csv file the kinematic of the agent over the capture
     """
     decoder = DecodedMessage()
-    f = open(filename, "r")
+    f = open(input_filename, "r")
     data = json.load(f)
     f.close()
 
@@ -22,7 +26,7 @@ def csv_conversion(filename: str, csv_filename: str, csv_interpolation: bool, st
     }
 
     if start_time:
-        data = filter_by_start_time(data, start_time)
+        data = utils.filter_by_start_time(data, start_time)
 
     last_update_pos = None
     i = 0
@@ -61,19 +65,19 @@ def csv_conversion(filename: str, csv_filename: str, csv_interpolation: bool, st
             continue
 
         if csv_interpolation and speed and last_update_pos:
-            if speed > SPEED_THRESHOLD and (d["timestamp"] - last_update_pos) / 1e3 > AGE_THRESHOLD:
+            if speed > utils.SPEED_THRESHOLD and (d["timestamp"] - last_update_pos) / 1e3 > utils.AGE_THRESHOLD:
                 pos_age = d["timestamp"] - last_update_pos
                 pos_age /= 1e3
-                interp_points = math.floor(pos_age / AGE_THRESHOLD)
+                interp_points = math.floor(pos_age / utils.AGE_THRESHOLD)
                 heading_diff = heading - last_heading if last_heading else 0
                 for j in range(0, int(interp_points)):
-                    t = last_update_pos + (j+1) * (AGE_THRESHOLD * 1e3)
+                    t = last_update_pos + (j+1) * (utils.AGE_THRESHOLD * 1e3)
                     delta_t = (t - last_update_pos) / 1e6
                     new_heading = last_heading + j * (heading_diff / interp_points)
                     delta_x = speed * delta_t * math.sin(new_heading)
                     delta_y = speed * delta_t * math.cos(new_heading)
-                    new_lat = lat + (delta_y / METERS_PER_DEGREE_LATITUDE)
-                    new_lon = lon + (delta_x / (METERS_PER_DEGREE_LATITUDE * math.cos(math.radians(lat))))
+                    new_lat = lat + (delta_y / utils.METERS_PER_DEGREE_LATITUDE)
+                    new_lon = lon + (delta_x / (utils.METERS_PER_DEGREE_LATITUDE * math.cos(math.radians(lat))))
                     interp_lat = new_lat
                     if not tmp_lat:
                         lat = interp_lat
